Refactor(sangjun/detailView): 게시판 상세 댓글API연결, 정규표현식 수정 #48
