+++ app/src/components/boards/read-detail/PostContainer.js
@@ -1,19 +1,12 @@
-import React, { useContext } from "react";
-import styled, { ThemeContext } from "styled-components/native";
+// import React, { useContext } from "react";
+// import styled, { ThemeContext } from "styled-components/native";
 
-import CommentContainer from "./comments/CommentContainer";
+// const Container = styled.View``;
 
-const Container = styled.View``;
+// const PostContainer = ({ getDateOrTime, detailViewInfo }) => {
+//   const theme = useContext(ThemeContext);
 
-const PostContainer = ({ getDateOrTime, detailViewInfo }) => {
-  const theme = useContext(ThemeContext);
+//   return <Container></Container>;
+// };
 
-  return (
-    <Container>
-      <>{detailViewInfo()}</>
-      <CommentContainer getDateOrTime={getDateOrTime} />
-    </Container>
-  );
-};
-
-export default PostContainer;
+// export default PostContainer;
