+++ app/src/screens/boards/read-all/FreeBoard.js
@@ -38,7 +38,6 @@ function FreeBoard({ navigation }) {
       );
       const json = await response.json();
       json.success ? setBoards(json.boards) : Alert.alert(json.msg);
-      console.log(json.boards);
     } catch (e) {
       Alert.alert("실패", e.message);
     } finally {
