+++ app/src/screens/boards/read-all/Market.js
@@ -41,7 +41,6 @@ function Board({ route, navigation }) {
         config
       );
       const json = await response.json();
-      console.log(json.boards);
       json.success ? setBoards(json.boards) : Alert.alert(json.msg);
     } catch (e) {
       Alert.alert("실패", e.message);
